fix timer timezone bug
