+++ crates/service/src/logic/workflow/workflow.rs
@@ -1,6 +1,5 @@
 use core::matches;
 use std::pin::Pin;
-use std::str::FromStr;
 
 use crate::logic::executor::ExecutorLogic;
 use crate::logic::job::JobLogic;
@@ -17,9 +16,8 @@ use crate::{
 use anyhow::{Result, anyhow};
 use automate::bridge::msg::UpdateJobParams;
 use automate::scheduler::types::{RunStatus, UploadFile};
-use chrono::{Local, Utc};
-use croner::Cron;
-use croner::parser::CronParser;
+use chrono::Local;
+
 use entity::{
     executor, instance, job, tag_resource, team, workflow, workflow_process, workflow_process_edge,
     workflow_process_node, workflow_process_node_task, workflow_timer, workflow_version,
@@ -1769,73 +1767,6 @@ impl<'a> WorkflowLogic<'a> {
             .exec(&self.ctx.db)
             .await?;
         Ok(ret.rows_affected)
-    }
-
-    pub fn check_timer_expr(timezone: &str, expr: &str) -> Result<Vec<String>> {
-        let parsed_expr = match CronParser::builder()
-            .seconds(croner::parser::Seconds::Required)
-            .dom_and_dow(true)
-            .build()
-            .parse(&expr)
-        {
-            Ok(_) => expr.to_string(),
-            Err(e1) => match english_to_cron::str_cron_syntax(expr) {
-                Ok(english_to_cron) => {
-                    if english_to_cron != expr {
-                        if english_to_cron == "0 * * * * ? *" {
-                            anyhow::bail!("failed parse {} to cron expr, {}", expr, e1.to_string())
-                        } else {
-                            // english-to-cron adds the year field which we can't put off (currently)
-                            let cron = english_to_cron
-                                .split(' ')
-                                .take(6)
-                                .collect::<Vec<_>>()
-                                .join(" ");
-                            cron
-                        }
-                    } else {
-                        expr.to_string()
-                    }
-                }
-                Err(e2) => {
-                    anyhow::bail!(
-                        "failed parse cron expr, 1.{}, 2.{}",
-                        e1.to_string(),
-                        e2.to_string()
-                    )
-                }
-            },
-        };
-
-        let parsed_cron = match Cron::from_str(&parsed_expr) {
-            Err(e) => anyhow::bail!("failed build cron, {}", e.to_string()),
-            Ok(v) => v,
-        };
-
-        let mut now = Local::now();
-        let mut next_exec_times: Vec<String> = vec![];
-
-        for _ in 0..10 {
-            let next_time = match parsed_cron.find_next_occurrence(&now, false) {
-                Err(e) => anyhow::bail!("failed find next execution time, {}", e.to_string()),
-                Ok(v) => {
-                    now = v.clone();
-                    match timezone {
-                        "local" => v
-                            .with_timezone(&Local)
-                            .format("%Y/%m/%d %H:%M:%S")
-                            .to_string(),
-                        "utc" | _ => v
-                            .with_timezone(&Utc)
-                            .format("%Y/%m/%d %H:%M:%S")
-                            .to_string(),
-                    }
-                }
-            };
-            next_exec_times.push(next_time);
-        }
-
-        Ok(next_exec_times)
     }
 
     pub async fn save_timer(
