+++ crates/utils/Cargo.toml
@@ -8,3 +8,7 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 anyhow.workspace = true
 serde_json.workspace = true
+tokio-cron-scheduler.workspace = true
+english-to-cron.workspace = true
+croner.workspace = true
+chrono.workspace = true
